Add a actor_names_by_tags() function which returns a subset of actors by tags.
This is needed in order to be able to filter a subset of actors by tags. The
load() function takes a list of tags, but loads actors which have any of them.
If we want to load actors which have all of them, we must load them using one
of tags and then filter them using the new actor_names_by_tags function.